Standardized InfiniteScroll to follow the data configs just like the InfiniteDataTable Structure.
