--- src/main/java/gwt/material/design/incubator/client/infinitescroll/data/LoadCallback.java
@@ -1,32 +1,0 @@
-/*
- * #%L
- * GwtMaterial
- * %%
- * Copyright (C) 2015 - 2018 GwtMaterialDesign
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-package gwt.material.design.incubator.client.infinitescroll.data;
-
-public interface LoadCallback<T> {
-    /**
-     * Should be called when async load success.
-     */
-    void onSuccess(LoadResult<T> loadResult);
-
-    /**
-     * Should be called when async load failed.
-     */
-    void onFailure(Throwable caught);
-}
