+++ src/main/java/gwt/material/design/incubator/client/infinitescroll/events/CompleteEvent.java
@@ -22,7 +22,6 @@ package gwt.material.design.incubator.client.infinitescroll.events;
 import com.google.gwt.event.shared.EventHandler;
 import com.google.gwt.event.shared.GwtEvent;
 import com.google.gwt.event.shared.HasHandlers;
-import gwt.material.design.incubator.client.infinitescroll.data.DataSource;
 //@formatter:off
 
 /**
