+++ src/main/resources/gwt/material/design/incubator/AddinsIncubator.gwt.xml
@@ -23,6 +23,7 @@
         "http://gwtproject.org/doctype/2.7.0/gwt-module.dtd">
 <module rename-to='AddinsIncubator'>
     <inherits name="gwt.material.design.addins.GwtMaterialAddinsBase"/>
+    <inherits name="gwt.material.design.GwtMaterialTableBasic" />
     <inherits name="com.google.gwt.json.JSON"/>
 
     <!-- Specify the app entry point class -->
